Extracted helper function to a separate file
