+++ client/src/Feed/Feed.tsx
@@ -1,5 +1,6 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, { Component } from 'react';
+import { timeDifference } from '../Helpers/helper';
 import './Feed.scss';
 
 type Post = {
@@ -58,36 +59,6 @@ class Feed extends Component<IProps, IState> {
         return date.toDateString();
     }
 
-    // https://stackoverflow.com/questions/6108819/javascript-timestamp-to-relative-time
-    timeDifference(previous: number) {
-        const msPerMinute = 60 * 1000;
-        const msPerHour = msPerMinute * 60;
-        const msPerDay = msPerHour * 24;
-        const msPerMonth = msPerDay * 30;
-        const msPerYear = msPerDay * 365;
-
-        const elapsed = Date.now() - previous;
-
-        if (elapsed < msPerMinute) {
-            return Math.round(elapsed / 1000) + ' seconds ago';
-        }
-        else if (elapsed < msPerHour) {
-            return Math.round(elapsed / msPerMinute) + ' minutes ago';
-        }
-        else if (elapsed < msPerDay) {
-            return Math.round(elapsed / msPerHour) + ' hours ago';
-        }
-        else if (elapsed < msPerMonth) {
-            return Math.round(elapsed / msPerDay) + ' days ago';
-        }
-        else if (elapsed < msPerYear) {
-            return Math.round(elapsed / msPerMonth) + ' months ago';
-        }
-        else {
-            return Math.round(elapsed / msPerYear) + ' years ago';
-        }
-    }
-
     render() {
         return (
             <div className="posts-list-wrapper">
@@ -97,7 +68,7 @@ class Feed extends Component<IProps, IState> {
                         this.state.posts.map((post, index) => {
                             return (
                                 <div key={index} className="post-item">
-                                    <div className="post-info">Posted by {post.author} {this.timeDifference(post.createdOn)}</div>
+                                    <div className="post-info">Posted by {post.author} {timeDifference(post.createdOn)}</div>
                                     <div className="post-content">{post.content}</div>
                                     <div className="post-comments-count">
                                         <button type="button" className="toggle-comments-button">
